feat: make multiple version order configurable
The order in the multiple versions drop-down on the episode/movie pages for libraries managed by the plugin is now configurable through the plugin settings, and by default will use the imported at date, but you can change the ordering as an _advanced setting_ if you feel like it.

Thanks to @GlawenC (Discord: <@707281950355881994>) for doing the research needed to implement this feature and for providing the basic implementation that was built upon in this commit.

Closes #46

Co-authored-by: Glawen <GlawenClatuc@proton.me>